requirements.txt
llama-index==0.10.30
llama-index-llms-google-genai
llama-index-embeddings-huggingface
sentence-transformers
llama-index-vector-stores-faiss
faiss-cpu
streamlit
python-dotenv
pypdf
google-generativeai